feat: add basic `Manager` structures and Thing management interface
This introduces a `Thing` inteface that managers should be able to
reliably consume to manage a given type of thing. Thing management
should be idempotent and generic wherever possible. Currently, to
implement the `Thing` interface, you must implement a `Manage` method --
this is the main method that a manager should interact with, and it
should idempotently get the given thing to the desired state (eg:
ensure a file exists with the specified contents with the correct
permissions/ownership).

The current intended design is to have a 1:1 relationship between thing
types and managers, where each manager manages a collection of related
things. This isn't strictly needed, and may change.