feat: add flag for setting log level and fix log level setting
Previously, the logic for setting the log level based on the config
value was in `init()` in the logging package. While the logging package
is read into main.go, the data is not available to viper until well into
main() when flags are set/parsed, config is read in, etc, long after
logging's `init()` func has run.

This yoinks the logic out of logging and moves it into main after the
viper initialization is done (as well as inclues a few fixes).