fix: prevent duplicate inventory entries for hosts and modules
Previously the logic where the host/module was appended into the new
list of items for the inventory would happen _inside_ the loop where
files for the relevant inventory object types happened. This means that,
for instance, a single host would get registered in the inventory
multiples -- once when parsing the `modules` file for the host, once
when parsing the `variables` file for the host, once when parsing the
`roles` file, etc.