feat: ensure that duplicate modules are filtered out
A module can be assigned to any number of roles, and can also be ad-hoc
assigned to a host. This means that it's possible to queue up in the
manager and trigger the same configuration script for a module multiple
times in the same run. Since scripts should be designed to be idempotent
this shouldn't necessarily be an issue, but it can be unnecessarily
wasteful/expensive on compute resources.