fix: stop `filterDuplicateModules()` from inserting empty mods on output
This commit feels very ironic, given the function in question.

Anyway, this was previously creating inserting an empty Module into the
slice because it was being initialized with `make()` to length of the
map, and then _appended_ to within the loop. Simplified to just init an
empty slice and retain the append logic.