fix!: improve variable sourcing, prevent duplicates
This previously was appending the variable onto the vars to keep for
each failed comparison of a blacklisted var in the list (so more vars
in that list == more duplication of vars). Switched to a flag that gets
tripped to better handle only adding the var to the list of keepers
once.