fix(template): recover from template panics, don't crash whole daemon
Mango uses's `template.Must()` which panics if a template has bad
syntax. However, bad syntax in a single script's template shouldn't
crash the entire daemon. This adds a recover() that allows for treating
bad templates as an error and continuing on with other execution, not
only allowing other modules with valid templates to run, but for the
daemon overall to continue operating.