feat: allow for templating variables files for both hosts and modules
The direct purpose I can think where this can be useful is for setting
OS-specific variables on a host. For instance, using a host's OS
metadata that's availabe during templating to set a host-level variable
for an install command that is templated to match the correct OS
version. There is a mockup of what this could look like in the included
test inventory changes.

Since variables files are sourced one time (when the manager is
reloading it's inventory), there is limited data available within
templating. When loading host variables, there is only system/runtime
metadata, no host/module variables. When sourcing variables files, only
host variables are available.