ci: duplicate testbox host in inv for new arch test container
Needing to literally duplicate/copy over an existing host to duplicate
it is frustrating.

I think I'm going to add a new `groups` folder into the inventory to
allow for defining a glob/regexp for matching hostnames against the
inventory. That way, you could have a single inventory definition for eg
`testbox-*` or similar.