feat!: `requires` file on modules in inventory for explicit ordering
This commit adds the ability to create a file called `requires` within a
module, that contains a new-line separated list of other modules that
must be run before this module.

Behind the scenes, this adds an edge to the DAG of modules that are
managed by the manager routine. When running all modules, the
topological sort ensures that there is a consistent ordering.

Example output from test inventory:
```
[root@testbox-arch /]# cat /tmp/mango-dependency-testing
/opt/mango/inventory/modules/test-reload
/opt/mango/inventory/modules/test-env-vars
/opt/mango/inventory/modules/test-requires-1
/opt/mango/inventory/modules/test-template
/opt/mango/inventory/modules/test-requires-2
/opt/mango/inventory/modules/test-requires-3
/opt/mango/inventory/modules/test-requires-4
/opt/mango/inventory/modules/test-requires-5
[root@testbox-arch /]#
```