fix: allow unset log level flag, default to info if unset
When the flag was unset, the previous switch case was failing to parse
and handle that case. This adds handling to default log level to `info`
if the flag is unset (as well as log a warning). Fixes:

```
~/go/src/github.com/tjhop/mango (main [ U ]) -> ./dist/mango_linux_amd64_v1/mango --inventory.path /dev/null
time=2024-03-09T14:24:52.649-05:00 level=ERROR msg="Failed to parse log level from flag" err="Unsupported log level" log_level=""
```

The default level is info so setting it explicitly is for the humans
reading this chunk of logic.