ref(inv): `GetVariablesForHost()` also get valid group/role variables
Previously, the `GetVariablesForHost()` method would only return the
host vars file path for the variables file assigned directly to the
host. It now also collects the variables file paths for groups that the
given host is a member of and roles that are valid for the host.

The `GetVariablesForSelf()` method has been refactored to internally
call `GetVariablesForHost()` for the running host.