fix: don't use `GetHost()` as form of enrollment check
If a host is enrolled via groups but not explicitly via it's own host
configs, then `GetHost()` will return that the host is not found,
because it doesn't exist.

This updates the places where `GetHost()` was treated as an enrollment
check and properly checks enrollment now.

Note: with this fix in place, variables from role files are now properly
parsing and being included in template renders/module runs.