feat: incorporate sprig template funcs, under the `sprig_` namespace
Sprig is the underlying template library used by helm (and lots of
others), so even though the project isn't actively developed anymore, it
still seems like a useful addition.

Functions are exposed under the `sprig_` namespace. For instance, the
`max` function in sprig is exposed as `sprig_max`.

NOTE: importantly, similar to helm, we are disabling the `env` and
`expandenv` template functions -- both for safety reasons as helm has
described, as well as the simple fact that mango sets/manages the
environment variables that are present for both the templating and
execution stages for a given script, which could yeild weird results.