fix(template): block device queue stats path update + error checking
The function will prepend the `/sys/block/` path in for us, so we don't
need to use the joined path here, just the regular device name returned
from the list function.

Also, somehow, I wrote this so that the error log would always log,
regardless of whether or not there was an error. It actually checks if
the function errored now. :facepalm: