feat(modules): allow loading user template files from `templates/*.tpl`
This adds support for pre-defining template files in
`$moduleName/templates/$templateFile.tpl`. The inventory will parse all
templates that match the glob `templates/*.tpl` within a module
directory into the inventory, which the manager will load and support to
the `templateScript()` function which is used to fully render template
files.

The test inventory's `test-template` module has a bit of mockup of how
it can be used.