fix!: properly check script exit codes, properly fail scripts if != 0
Previous code was not distinguishing between failures related to
shell/parsing libraries vs the scripts themselves returning non-zero
exit codes and failing. This reworks the `shell.Run()` func to actually
return the scripts exit code, as well as updates the callsites where
it's used to better track/log the info for users.