fix(manager): don't shadow module test script exit code
This was redeclaring the `testRC` var defined in the outer scope. The
`err` var already exists in this scope from the prior calls, so both of
these vars can be updated to just assignment, no declaration needed (and
thus no shadowing). This preserves the test script's exit code, rather
than always returning 0 because it was in the wrong scope.