fix(build): fix env var usage in packaged systemd unit file
From docs
(https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html#Command%20lines):

> Basic environment variable substitution is supported. Use "${FOO}" as
part of a word, or as a word of its own, on the command line, in which
case it will be erased and replaced by the exact value of the
environment variable (if any) including all whitespace it contains,
always resulting in exactly a single argument. Use "$FOO" as a separate
word on the command line, in which case it will be replaced by the value
of the environment variable split at whitespace, resulting in zero or
more arguments. For this type of expansion, quotes are respected when
splitting into words, and afterwards removed.

Using curly brackets was causing this to be interpeted by mango as a
single massive command line arg. We want splitting here.