feat(inv): add support for user defined templates to host, role, group
Similar to how it's supported for modules -- there's a `templates/`
folder with `*.tpl` files that get parsed.

The inventory will store the paths to valid template files in each
component. These will be used by the manager in the templating runs of
all files that support templating, not just modules.