fix(logging): don't duplicate manager group key
I'd never realized this before, but `.With()` is append only, and so
re-adding a set of kvs under an existing group adds a second group with
the new entries -- it does not combine them in the json object or add
the keys to the existing group like I had for some reason thought.

This fixes that by only setting the kvs for that group once.