chore: add normalizeStringFlag helper func, improve flag handling
intended for use with flags that accept a custom string input, but which
are used/mapped to a specific value. there are lots of flags which do
not need this normalize because they are _intended_ to be accepted and
used literally (inventory path, hostname, reload interval, etc). but the
flags here accept a custom string input which maps to a log level or log
format, respectively. adding this func standardizes that normalization.