chore: improve logging around unset log level flags
previously, if the log level wasn't set via flag, this would fall
through to the default path, which contains strong wording such as
"Failed to parse log level from flag", and setting an `err` field on a
warning level log that calls the value from the flag "unsupported". the
flag is indeed optional, and it feels weird to yell about it. this
addresses that issue by improving logging and handing of the unset flag
case.