ref!: upgrade to sprout 0.5.1, template behavior has changed
As sprout has adopted it's new registry system, this refactors the
template layer to use the new handler system. For consistency, the `env`
registry is excluded for security reasons as was previously done, and
all other registries (except the `backward` compat registry) are enabled
and built into the templating engine in mango.

As sprout has also adopted it's updated/improved versions of templating
functions, there are also a few changes in behavior that required test
changes and were noticed, but there may be others. In particular:

- `replace` replaces _all_ occurences by default now, rather than just
  the first
- `regexReplaceAll` better handles regexs it seems and made me notice
  the regex in the test template wasn't really doing what I intended, so
there's that as well