test: add intentionally unenrolled host, rename existing hosts
To better test how things behave when a host is unenrolled, it's helpful
to have one. This renames/modifies references to the existing test boxes
and includes a `-mango-` in the name to indicate it is intended to be
enrolled, and the box intended to be unenrolled gets a `-unenrolled-`
in it's config names/host name.